requirements.txt
requests==2.28.2
python-telegram-bot==20.0
asyncio==3.4.3
